Reintegrated intervening extrema.
